+++ telescope/src/HomepageComponets/NavbarTopComponents.css
@@ -1,0 +1,43 @@
+/* NavbarTopComponents.css */
+
+
+
+/* Style the container for LocationSelect and SearchBar */
+.container {
+    display: flex;
+    align-items: center;
+    background-color: #e0dcdc; /* Background color */
+    padding: 10px; /* Padding for spacing */
+    border-radius: 8px; /* Rounded corners */
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Box shadow for a subtle lift */
+    transition: all 0.3s; /* Smooth transition on hover */
+  }
+  
+  /* Style the LocationSelect and SearchBar */
+  .items {
+    display: flex;
+    padding: 26px; /* Add padding */
+    transition: all 0.3s; /* Smooth transition on hover */
+  }
+  
+  /* Add hover effect */
+  .items:hover {
+    background-color: #fff; /* Change background color on hover */
+    transform: translateY(-2px); /* Slight lift on hover */
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Increased box shadow on hover */
+  }
+  
+  /* Media query for responsiveness */
+  @media (max-width: 768px) {
+    .container {
+      flex-direction: column; /* Stack components on small screens */
+    }
+  }
+
+  .navbar-logo{
+    width: 50px;
+    height: 50px;
+    display: flex;
+  }
+  
+
