+++ telescope/src/HomepageComponets/SearchBar.jsx
@@ -1,0 +1,24 @@
+// SearchBar.jsx
+import React, { useState } from 'react';
+
+const SearchBar = ({ onSearch }) => {
+  const [searchText, setSearchText] = useState('');
+
+  const handleSearch = () => {
+    onSearch(searchText); // Pass the search text to the parent component
+  };
+  
+  return (
+    <div className="search-bar">
+      <input
+        type="text"
+        placeholder="Search..."
+        value={searchText}
+        onChange={(e) => setSearchText(e.target.value)}
+      />
+      <button onClick={handleSearch}>Search</button>
+    </div>
+  );
+};
+
+export default SearchBar;
